requirements.txt
streamlit==1.52.2
aiofiles==25.1.0
ollama==0.6.1
openai==2.15.0
pyyaml==6.0.1
fastapi==0.128.0
langchain==1.2.3
langchain-community==0.4.1
langchain-core==1.2.7
langchain-ollama==1.0.1
langchain-openai==1.1.7
openpyxl==3.1.5
pandas==2.3.3
pydantic==2.12.5
python-dotenv==1.0.1
python-multipart==0.0.21
requests==2.32.5
tiktoken==0.12.0
uvicorn==0.40.0
